refactor: resultArea 초기화 함수 분리
