+++ src/controller/resultAreaController.js
@@ -1,0 +1,8 @@
+export const clearResultTable = () => {
+	const tableContainer = document.querySelector('button + div > table');
+	const tableTitle = document.querySelector('button + div > p');
+	if (tableContainer) {
+		tableContainer.remove();
+		tableTitle.remove();
+	}
+};
