+++ src/utils/elementUtils.js
@@ -32,12 +32,3 @@ export const appendChilds = (parent, childs) => {
 		parent.appendChild(child);
 	});
 };
-
-export const clearResultTable = () => {
-	const tableContainer = document.querySelector('button + div > table');
-	const tableTitle = document.querySelector('button + div > p');
-	if (tableContainer) {
-		tableContainer.remove();
-		tableTitle.remove();
-	}
-};
